refactor: move file logic to services
